fix: 산책 경로 추천 필터링 (#118)
